# Optimising the minimum function for speed : revert to min(a,b)