name              old time/op    new time/op    delta
Simple/ASCII-4       471ns ± 1%     393ns ± 2%  -16.44%  (p=0.000 n=25+28)
Simple/French-4      842ns ± 2%     719ns ± 1%  -14.66%  (p=0.000 n=28+30)
Simple/Nordic-4     1.67µs ± 1%    1.43µs ± 0%  -14.59%  (p=0.000 n=28+28)
Simple/Tibetan-4    1.31µs ± 1%    1.15µs ± 1%  -12.15%  (p=0.000 n=27+29)
[Geo mean]           966ns          826ns       -14.48%

name              old alloc/op   new alloc/op   delta
Simple/ASCII-4       96.0B ± 0%     96.0B ± 0%     ~     (all equal)
Simple/French-4       128B ± 0%      128B ± 0%     ~     (all equal)
Simple/Nordic-4       192B ± 0%      192B ± 0%     ~     (all equal)
Simple/Tibetan-4      144B ± 0%      144B ± 0%     ~     (all equal)
[Geo mean]            136B           136B        +0.00%

name              old allocs/op  new allocs/op  delta
Simple/ASCII-4        1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Simple/French-4       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Simple/Nordic-4       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Simple/Tibetan-4      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
[Geo mean]            1.00           1.00        +0.00%