@psadac 2018-09-11
Hello Agniva,
changing the function min(a, b, c) to min(a, b) seems to have a great impact on ComputeDistance() performance.
I am a newbie in Go, so there may be something wrong with what i have done : 
`for i in {0..29}; do make bench >> [old,new].txt; done`
then : 
`benchstat -geomean old.txt new.txt`