@agnivade 2018-09-12
This is really surprising ! I checked it, and I see it too.

FYI, you do not need to do this - `for i in {0..29}; do make bench >> [old,new].txt; done`. You can specify the `-count` parameter in `go test` itself. That is more idiomatic.

I would like to see the assembly output and see why we are seeing this gain. If you have some time, you can get the assembly output and paste it here. Or I would do it myself in some time.