@psadac 2018-09-12
I have generated the assembly output  : 
`go tool compile -S levenshtein.go > [old,new].s`
and compared the relevant portion of the previous version : 
>         cmpq    R13, AX
>         jcc     659
>         movq    -8(CX)(R10*8), R11
>         cmpq    R10, AX
>         jcc     659
>         incq    R11
>         leaq    1(R8), R12
>         movq    (CX)(R10*8), R14
>         incq    R14
>         cmpq    R11, R12
>         jge     445
>         cmpq    R11, R14
>         jlt     383
>         movq    R14, R11
>         jmp     383
>         cmpq    R12, R14
>         jge     440
>         movq    R12, R11
>         jmp     383
with the new one (using the [Compiler Explorer](https://godbolt.org/)) : 
>         cmpq    R13, AX
>         jcc     651
>         movq    -8(CX)(R10*8), R11
>         incq    R11
>         leaq    1(R8), R12
>         cmpq    R11, R12
>         cmovqlt R11, R12
>         cmpq    R10, AX
>         jcc     651
>         movq    (CX)(R10*8), R14
>         leaq    1(R14), R11
>         cmpq    R12, R11
>         cmovqlt R12, R11
>         jmp     383
my understanding in assembly is basic but i think that : 
-  `min()` functions are inlined in both versions
-  SSA compiler has generated optimised `cmovqlt` instructions in the new version (`cmovqlt` is present in go source code file `.../ssa/rewriteAMD64.go`)
- there are less jumps in the newer version (3 vs 7) so the flow in the instruction pipeline of the processor is probably optimised

Thank you for help, i didnt know the `-count` parameter, and feel free correct me if i made incorrect assumptions.